Refactor resources.py into resources package with apis grouped into separate files. Create blueprint for each group and register in app.py.
